requirements.txt
paho-mqtt==1.3.0
python-dotenv==0.6.4